cleanup country selector
